+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## master (unreleased)
 
-- #2 Make plugin for chai
+- [#2](https://github.com/szikszail/date-x/issues/2) Make plugin for chai
 
 ## 0.2.0